+++ test/solve_test.cpp
@@ -3,21 +3,38 @@
  * solving Laplacian systems.
  */
 #include <stdexcept>
-#include <iostream>
 #include <gtest/gtest.h>
 #include <graph.h>
 #include <solve.h>
 
-TEST(SolveTest, JacobiIteration) {
-  // Create a small cycle graph
+TEST(SolveTest, JacobiIterationConvergenceFailure) {
+  // Create a small cycle graph - Jacobi iteration cannot solve this system.
   stag_int n = 5;
   stag::Graph testGraph = stag::cycle_graph(n);
 
   // Construct the target vector
   DenseVec b(n);
-  b << 2, 4, 1, 6, 4;//, 5, 3, 2, 1, 3;
+  b << 2, 4, 1, 6, 4;
 
   // Solve a linear system
-  DenseVec soln = stag::solve_laplacian(&testGraph, b, 0.1);
+  EXPECT_THROW(stag::solve_laplacian_jacobi(&testGraph, b, 0.1),
+               stag::ConvergenceError);
 }
 
+TEST(SolveTest, JacobiIteration) {
+  // Create a small second difference graph
+  stag_int n = 5;
+  stag::Graph testGraph = stag::second_difference_graph(n);
+
+  // Construct the target vector
+  DenseVec b(n);
+  b << 2, 4, 1, 6, 4;
+
+  // Solve the linear system
+  DenseVec x_hat = stag::solve_laplacian_jacobi(&testGraph, b, 0.1);
+
+  // Compare with the expected answer
+  DenseVec x(n);
+  x << (15./2), 13, (29./2), 15, (19./2);
+  EXPECT_TRUE(x_hat.isApprox(x, 0.01));
+}